+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - Begin the texture update.
 - Add specific blood, DNA, and egg types to a creative tab.
     - This means that Honk types are now synchronised to the client from a dedicated server.
+- Add a processing sound for the DNA Combinator.
 
 [Full Changelog](https://github.com/JamCoreModding/Honk/compare/0.1.0-beta.8...0.1.0-beta.9)